Fixed adding force to non dof segment

+++ examples/WrappingObjectsExample.cpp
@@ -28,7 +28,7 @@ int main()
 
     // Set all muscles to half of their maximal activation
     std::vector<std::shared_ptr<internal_forces::muscles::State>> states;
-    for (unsigned int i=0; i<model.nbMuscles(); ++i) {
+    for (size_t i=0; i<model.nbMuscles(); ++i) {
         states.push_back(std::make_shared<internal_forces::muscles::State>(0, 0.5));
     }
