+++ include/InternalForces/Actuators/ActuatorSigmoidGauss3p.h
@@ -54,7 +54,7 @@ public:
         const utils::Scalar& offset,
         const utils::Scalar& r,
         const utils::Scalar& qopt,
-        unsigned int dofIdx);
+        size_t dofIdx);
 
     ///
     /// \brief Construct Gauss3p actuator
@@ -74,7 +74,7 @@ public:
         const utils::Scalar& offset,
         const utils::Scalar& r,
         const utils::Scalar& qopt,
-        unsigned int dofIdx,
+        size_t dofIdx,
         const utils::String &jointName);
 
     ///
